requirements.txt
streamlit
google-generativeai
PyPDF2
fpdf
python-docx